Refactor: Promote numerals and split main.rs
- Moved `src/experimental/numerals.rs` to `src/parser/numerals.rs` and deleted `experimental` module.
- Split `src/main.rs` into `src/tools/cli.rs`, `src/tools/repl.rs`, and `src/tools/runner.rs`.
- `main.rs` is now a thin entry point.
- Updated `src/tools/mod.rs` to export new modules.
- Updated `src/parser/builder.rs` and `src/parser/mod.rs` to reflect numeral changes.
- Added corresponding tests to new tool modules.

This aligns with Razor's philosophy of flattening abstractions and removing dead/experimental code.
