+++ src/lib.rs
@@ -48,7 +48,6 @@
 pub mod ast;
 pub mod codegen;
 pub mod errors;
-pub mod experimental;
 pub mod grammar;
 pub mod morphology;
 pub mod parser;
