+++ src/parser/mod.rs
@@ -23,6 +23,7 @@
 //! * Provide better error messages during the conversion phase.
 
 pub(crate) mod builder;
+pub mod numerals;
 
 use crate::ast::Program;
 use crate::errors::GlossaError;
