+++ src/tools/cli.rs
@@ -1,0 +1,51 @@
+//! Command-line interface definition
+
+use clap::{Parser, Subcommand};
+use std::path::PathBuf;
+
+#[derive(Parser)]
+#[command(name = "glossa")]
+#[command(about = "ΓΛΩΣΣΑ - Ancient Greek morphology as programming semantics")]
+#[command(version)]
+pub struct Cli {
+    #[command(subcommand)]
+    pub command: Option<Commands>,
+
+    /// Run a .γλ file directly (default action)
+    #[arg(value_name = "FILE")]
+    pub file: Option<PathBuf>,
+}
+
+#[derive(Subcommand)]
+pub enum Commands {
+    /// Run a .γλ file (default)
+    Run {
+        /// Input file (.γλ)
+        input: PathBuf,
+    },
+
+    /// Compile a .γλ file to Rust source
+    Build {
+        /// Input file (.γλ)
+        input: PathBuf,
+
+        /// Output file (.rs)
+        #[arg(short, long)]
+        output: Option<PathBuf>,
+    },
+
+    /// Check a .γλ file without running
+    Check {
+        /// Input file (.γλ)
+        input: PathBuf,
+    },
+
+    /// Highlight a .γλ file with semantic colors
+    Highlight {
+        /// Input file (.γλ)
+        input: PathBuf,
+    },
+
+    /// Start the interactive REPL
+    Repl,
+}
