+++ src/tools/mod.rs
@@ -1,1 +1,4 @@
+pub mod cli;
 pub mod highlight;
+pub mod repl;
+pub mod runner;
